[MWM-7] Connect to Postgres via Hibernate (#15)
* MWM-7 Makes a whole host of adjustments to make the app work against a real PostgreSQL database. Changes config files, installs the postgres driver, changes the hibernate dialect, removes all double types in models (double is not supported in postgres) and adds a try/catch around multiple joins, some of which will return no results

* MWM-7 runs keycloak in the docker-compose file, adds a seed script to automatically create the monmon-play pg database, adds a post_run_boh.sh script that sets up the Monmon keycloak realm

+++ app/models/Balance.java
@@ -11,7 +11,7 @@ public class Balance {
     @GeneratedValue(strategy = GenerationType.AUTO)
     public Long id;
     public Long timestamp; // this makes this time series data, for graphs and junk
-    public Double value;
+    public Float value;
 
     @ManyToOne
     @JoinColumn(name="account_id")
@@ -45,11 +45,11 @@ public class Balance {
         this.timestamp = timestamp;
     }
 
-    public Double getValue() {
+    public Float getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(Float value) {
         this.value = value;
     }
